Let users click-drag notes in the pattern editor
Previously, clicking and dragging on the pattern editor would add a new
note to every cell that the mouse passes over. That made it a bit
diffult to make a lot of incremental changes, since there was no way to
drag notes around.

By default, clicking and dragging on a note will move it around.  Hold
SHIFT to get the previous behavior of adding a new note to every cell.